feat: Clean up `$HOME`
Keep `$HOME` completely free of `zgenom`, `zsh-quickstart-kit`,
`iterm2_shell_integration.zsh`, and all standard `zsh` shell files.

This only requires `$ZDOTDIR` to be set on shell login (usually in
`~/.zshenv` or `/etc/zshenv`).

Files and paths:

- `$ZDOTDIR`:
  - `.zshrc`, `.zgen-setup`, `.zsh_aliases`, `.zsh_functions`
  - `zshrc.d/`
  - `zshrc.pre-plugins.d/`
  - `.zqs/` (stores `zsh-quickstart-kit` settings)
  - `.iterm2.zsh` (shell integration script)
  - `.zhistory` (previously `.zsh_history`)
  - `.zcomp*` (completion files, previously `~/.zsh/complete`)
- `~/.local/share/`:
  - `zqs-zgenom/` (`zgenom` repo, previously `~/zgenom/`)
  - `zgenom/` (plugins, init.zsh, etc., previously `~/.zgenom/`)

The following can't(?) be moved (handled by respective programs):

- `~/.fzf/`
- `~/.fzf.zsh`
- `~/.p10k.zsh`

PLEASE NOTE: This is only a proof of concept! Much of it is hardcoded,
only tested on a fresh install of `zsh` with `homebrew` on macOS 13,
and barely tested in general.

BREAKING CHANGES:

- Causes even _messier_ $HOME if `$ZDOTDIR` is set to `$HOME`.
- Probably won't work at all if `$ZDOTDIR` is unset.
- Probably won't work if `zsh-quickstart-kit` has already been installed
- Requires manually moving `~/.iterm2_shell_integration.zsh`:
  - `$ZDOTDIR/.iterm2.zsh`